reworked with p-queue, currently n=1 works

+++ disk_class.h
@@ -4,35 +4,35 @@
 #ifndef DISK_CLASS
 #define DISK_CLASS
 
-typedef std::pair <float, float> vec;
+typedef std::pair <double, double> vec;
 
 class Disk {
 	public:
 		Disk();
 		Disk(Uint32 pixel);
-		Disk(vec c, float r, Uint32 pixel);
-		Disk(vec c, float r, vec v, Uint32 pixel);
+		Disk(vec c, double r, Uint32 pixel);
+		Disk(vec c, double r, vec v, Uint32 pixel);
 		void set_c(vec cin);
 		void set_v(vec vin);
-		void set_r(float radius);
+		void set_r(double radius);
 		void set_color(Uint32 pixel);
 		
 		vec get_c();
 		vec get_v();
-		float get_r();
+		double get_r();
 		
 		void render(SDL_Surface *surface);
-		void move(SDL_Surface *surface, float t);
+		void move(SDL_Surface *surface, double t);
 
 	private:
-		void render_backend(SDL_Surface *surface, float r, Uint32 pixel);
+		void render_backend(SDL_Surface *surface, double r, Uint32 pixel);
 		void set_pixel(SDL_Surface *surface, int x, int y, Uint32 pixel);
-		bool resolve_wall_xcollisions(float w, float t);
-		bool resolve_wall_ycollisions(float h, float t);
+		bool resolve_wall_xcollisions(double w, double t);
+		bool resolve_wall_ycollisions(double h, double t);
 		Uint32 color;
 		vec c;
 		vec v;
-		float r;
+		double r;
 };
 
 #endif /* DISK_CLASS */
